+++ src/GafferModule/SerialisationBinding.cpp
@@ -73,6 +73,18 @@ std::string childIdentifier( const Serialisation &serialisation, const std::stri
 	return serialisation.childIdentifier( parentIdentifier, child );
 }
 
+std::string objectToBase64Wrapper( const IECore::Object *object )
+{
+	IECorePython::ScopedGILRelease gilRelease;
+	return Serialisation::objectToBase64( object );
+}
+
+IECore::ObjectPtr objectFromBase64Wrapper( const std::string &base64String )
+{
+	IECorePython::ScopedGILRelease gilRelease;
+	return Serialisation::objectFromBase64( base64String );
+}
+
 } // namespace
 
 void GafferModule::bindSerialisation()
@@ -97,6 +109,10 @@ void GafferModule::bindSerialisation()
 		.staticmethod( "modulePath" )
 		.def( "classPath", (std::string (*)( object & ))&Serialisation::classPath )
 		.staticmethod( "classPath" )
+		.def( "objectToBase64", &objectToBase64Wrapper )
+		.staticmethod( "objectToBase64" )
+		.def( "objectFromBase64", &objectFromBase64Wrapper )
+		.staticmethod( "objectFromBase64" )
 		.def( "registerSerialiser", &Serialisation::registerSerialiser )
 		.staticmethod( "registerSerialiser" )
 		.def( "acquireSerialiser", &Serialisation::acquireSerialiser, return_value_policy<reference_existing_object>() )