+++ Changes.md
@@ -13,6 +13,7 @@ Fixes
 API
 ---
 
+- Serialisation : Added `objectToBase64()` and `objectFromBase64()` methods to provide base64 encoding and decoding for all `IECore::Objects`.
 - NumericWidget : Fixed bug that caused `editingFinished` to be called with the wrong `reason` when the widget was left with an invalid value. `Invalid` is now passed in these cases.
 
 0.58.4.0 (relative to 0.58.3.2)